+++ lib/mix/tasks/live_vue.install.ex
@@ -633,11 +633,9 @@ defmodule Mix.Tasks.LiveVue.Install do
     defp server_js_content do
       """
       import components from "../vue"
-      import { getRender, loadManifest } from "live_vue/server"
+      import manifest from "live_vue/ssrManifest"
+      import { getRender } from "live_vue/server"
 
-      // present only in prod build. Returns empty obj if doesn't exist
-      // used to render preload links
-      const manifest = loadManifest("../priv/static/.vite/ssr-manifest.json")
       export const render = getRender(components, manifest)
       """
     end
@@ -718,7 +716,7 @@ defmodule Mix.Tasks.LiveVue.Install do
             String.replace(
               content,
               ~s("phoenix_vite.npm vite build"),
-              ~s("phoenix_vite.npm vite build --manifest --emptyOutDir true", "phoenix_vite.npm vite build --ssrManifest --emptyOutDir false --ssr js/server.js --outDir ../priv/static")
+              ~s("phoenix_vite.npm vite build --manifest --ssrManifest --emptyOutDir true", "phoenix_vite.npm vite build --emptyOutDir false --ssr js/server.js --outDir ../priv/static")
             )
           end
         end)